tests: hoist minimal_conf_noscreen() to the top level tests
we'll use this in the next patch for testing serial number ordering.
