+++ test/widgets/conftest.py
@@ -5,9 +5,6 @@ import subprocess
 
 import pytest
 
-import libqtile.config
-import libqtile.confreader
-import libqtile.layout
 from libqtile.bar import Bar
 from libqtile.widget.base import ORIENTATION_HORIZONTAL
 
@@ -71,25 +68,6 @@ def fake_qtile():
                 loop.close()
 
     return FakeQtile()
-
-
-# Fixture that defines a minimal configurations for testing widgets.
-# When used in a test, the function needs to receive a list of screens
-# (including bar and widgets) as an argument. This config can then be
-# passed to the manager to start.
-@pytest.fixture(scope="function")
-def minimal_conf_noscreen():
-    class MinimalConf(libqtile.confreader.Config):
-        auto_fullscreen = False
-        keys = []
-        mouse = []
-        groups = [libqtile.config.Group("a"), libqtile.config.Group("b")]
-        layouts = [libqtile.layout.stack.Stack(num_stacks=1)]
-        floating_layout = libqtile.resources.default_config.floating_layout
-        screens = []
-        reconfigure_screens = False
-
-    return MinimalConf
 
 
 @pytest.fixture(scope="function")
